Added some board functionailty but tbh might be doing too much. Need to learn more about how the learning actually happends in order to design the classes / data structures better
