+++ transformCard.py
@@ -1,0 +1,6 @@
+def cardToBoard(card):
+    #View _GameBoard.py to learn more about the board structure
+    #Return an int value for the index of the game board
+    #TODO
+
+    return 1
